3164: Added media embed code to API response
