+++ config/services.yaml
@@ -39,7 +39,7 @@ services:
               templates: #'%env(json:MEDIA_PROCESSOR_TEMPLATES)%'
                 -
                     pattern: "~^https://media.videotool.dk\\?vn=(?P<id>[^&#]+)&is_audio=1$~"
-                    template: |
+                    template: >-
                         <div style="position: relative; padding-bottom: 26.25%%; height: 0;">
                         <iframe aria-label="{{ title }}"
                                 type="text/html"
@@ -52,7 +52,7 @@ services:
 
                 -
                     pattern: "~^https://media.videotool.dk\\?vn=(?P<id>[^&#]+)$~"
-                    template: |
+                    template: >-
                         <div style="position: relative; padding-bottom: 56.25%%; height: 0;">
                         <iframe aria-label="{{ title }}"
                                 type="text/html"
