+++ src/Service/MediaProcessorInterface.php
@@ -9,5 +9,5 @@ interface MediaProcessorInterface
     /**
      * Get embed code.
      */
-    public function getEmbedCode(PointOfInterest $point): string;
+    public function getEmbedCode(PointOfInterest $entity, string $property = 'mediaUrl'): string;
 }
